04_RAG_AND_AI_AGENTS/INTRO_TO_GENERATIVE_AI/01_BASIC_RAG/STAGE_3_APP_lvl2/INGESTION/requirement.txt
pymilvus
google-genai
numpy
pydantic
python-dotenv
pandas